pkg/controller: abstract getManagedKey to be used in other ctrls
