+++ pkg/controller/kubelet-config/kubelet_config_features.go
@@ -20,7 +20,6 @@ import (
 	kubeletconfigv1beta1 "k8s.io/kubelet/config/v1beta1"
 
 	ctrlcommon "github.com/openshift/machine-config-operator/pkg/controller/common"
-	mtmpl "github.com/openshift/machine-config-operator/pkg/controller/template"
 	"github.com/openshift/machine-config-operator/pkg/version"
 )
 
@@ -94,7 +93,7 @@ func (ctrl *Controller) syncFeatureHandler(key string) error {
 		isNotFound := errors.IsNotFound(err)
 		if isNotFound {
 			ignConfig := ctrlcommon.NewIgnConfig()
-			mc, err = mtmpl.MachineConfigFromIgnConfig(role, managedKey, ignConfig)
+			mc, err = ctrlcommon.MachineConfigFromIgnConfig(role, managedKey, ignConfig)
 			if err != nil {
 				return err
 			}
